requirements.txt
asgiref==3.8.1
Django==5.1.2
django-unfold==0.40.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
